+++ src/app.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import U2k from './pages/4u2k_page'
-import B2B from './b2b_page'
-import Main from './main_page'
+import B2B from './pages/b2b_page'
+import Main from './pages/main_page'
 import ABC from './abc'
 import DoUHearMe from './douhearme'
 import Games from './games'
@@ -11,7 +11,7 @@ import { ClientContactForm } from './components/client-contact-form'
 const App = () => {
     return (
         <body>
-            < Price />
+            < Main />
         </body>
     );
 };
