+++ src/components/invite-button/index.ts
@@ -1,0 +1,1 @@
+export { InviteButton } from './invite-button';
