+++ src/components/schedule-info/index.ts
@@ -1,0 +1,1 @@
+export { ScheduleInfo } from './schedule-info';
