+++ src/components/who-we-are-div/who-we-are-div.tsx
@@ -1,0 +1,29 @@
+import React from "react";
+
+export const WhoWeAreDiv = () => {
+    return (
+        <div className="whoweare_div_div">
+            <div className="whoweare_div_div_p">
+                <h2 className="whoweare_div_div_number">1</h2>
+                <p className="whoweare_div_div_info">Мы создаем пространство, где каждый чувствует себя <span
+                    className="red">комфортно</span> и готов общаться.</p>
+            </div>
+            <div className="whoweare_div_div_p">
+                <h2 className="whoweare_div_div_number">2</h2>
+                <p className="whoweare_div_div_info">Мы проводим встречи в разных местах, чтобы каждый мог выбрать
+                    то, что ему больше <span className="red">нравится.</span></p>
+            </div>
+            <div className="whoweare_div_div_p">
+                <h2 className="whoweare_div_div_number">3</h2>
+                <p className="whoweare_div_div_info">Наши встречи ориентированы на <span className="red">практику </span>
+                        английского языка, а не занудную теорию.</p>
+            </div>
+            <div className="whoweare_div_div_p">
+                <h2 className="whoweare_div_div_number">4</h2>
+                <p className="whoweare_div_div_info">Здесь ты найдешь тех самых <span className="red">интересных </span>
+                    людей с которыми так здорово проводить время вместе.</p>
+            </div>
+            <h2 className="whoweare_div_div_header"><span className="red">HEY,</span> LET'S GO</h2>
+        </div>
+    );
+}
